+++ Documentation/Changelog.md
@@ -113,6 +113,7 @@ For more details, please refer to:
 
 For more details, please refer to:
 
+* [MariaDB MaxScale 22.08.6 Release Notes](Release-Notes/MaxScale-22.08.6-Release-Notes.md)
 * [MariaDB MaxScale 22.08.5 Release Notes](Release-Notes/MaxScale-22.08.5-Release-Notes.md)
 * [MariaDB MaxScale 22.08.4 Release Notes](Release-Notes/MaxScale-22.08.4-Release-Notes.md)
 * [MariaDB MaxScale 22.08.3 Release Notes](Release-Notes/MaxScale-22.08.3-Release-Notes.md)